New option -(no-)fix-orientation
When the mesh is read in MFEM, the orientation of the cells can be fixed,
and it's fixed by default. That can be changed with this new option '-no-fix-orientation'.
In the latter case the mesh will be used as is. In some cases it gives more accurate simulation
results at receivers.