New approach to compute R matrices on parallel grids
Now, each process has a map between the coarse cell ID and a list of fine cells belonging to that coarse cell.
Based on this map, each process takes its coarse cell, creates a Gmsh mesh file for this fine grid, reads it
to create a Mesh object and solves the local eigenproblems on that mesh.